+++ src/typealiases/InlineQueryResult/sentWebAppMessage.ts
@@ -1,0 +1,3 @@
+export type SentWebAppMessage = {
+    inline_message_id?: string //Identifier of the sent inline message. Available only if there is an inline keyboard attached to the message.
+}