+++ src/typealiases/webAppInfo.ts
@@ -1,0 +1,3 @@
+export type WebAppInfo = {
+    url: string //An HTTPS URL of a Web App to be opened with additional data as specified in Initializing Web Apps
+}